git-svn-id: svn+ssh://src.apple.com/svn/cups/easysw/current@3015 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ filter/error.c
@@ -78,7 +78,7 @@ _cupsRasterAddError(const char *f,	/* I - Printf-style error message */
   if (bytes >= sizeof(s))
     return;
 
-  if (bytes > (buf->end - buf->current))
+  if (bytes > (size_t)(buf->end - buf->current))
   {
    /*
     * Allocate more memory...
